separate incoming/outgoing db impl
